Assignment submission

+++ JuliaGPU/session1/Animesh/q2.jl
@@ -3,10 +3,10 @@ if n>0
   println(n)
   while(n!=1)
     if n%2==0
-       global n = n/2
+       global n = Int(n/2)
        println(n)
     else
-       global n = (3*n)+1
+       global n = Int((3*n)+1)
        println(n)
     end
   end
